Add -Wall at compilation and remove all warnings

+++ src/ai/strategy/Strategy.cpp
@@ -14,7 +14,7 @@ Strategy::Strategy():
  */
 void Strategy::update_player_ids(){
     this->player_ids.clear();
-    for( int i=0; i<robot_ids.size(); i++ ){
+    for( unsigned int i=0; i<robot_ids.size(); i++ ){
         if( goalie_id == robot_ids[i] ){
             robot_ids[i] = robot_ids[robot_ids.size()-1];
             robot_ids[robot_ids.size()-1] = goalie_id;
@@ -47,12 +47,12 @@ const std::vector<int> & Strategy::get_player_ids() const {
 }
 
 int Strategy::robot_id( int id ) const {
-    assert( 0<=id and id<robot_ids.size() ); // Whent that line make an assertion, that means, you don't have updated the min_robots() implementation inside your strategy code.
+    assert( 0<=id and static_cast<unsigned int>(id)<robot_ids.size() ); // Whent that line make an assertion, that means, you don't have updated the min_robots() implementation inside your strategy code.
     return robot_ids[id];
 }
 
 int Strategy::player_id( int id ) const {
-    assert( 0<=id and id<player_ids.size() ); // Whent that line make an assertion, that means, you don't have updated the min_robots() implementation inside your strategy code.
+    assert( 0<=id and static_cast<unsigned int>(id)<player_ids.size() ); // Whent that line make an assertion, that means, you don't have updated the min_robots() implementation inside your strategy code.
     return player_ids[id];
 }
